dataTableOutput and bug fix
Changed to dataTableOutput to be searchable and fixed an issue with not being able to search capabilities with spaces. 

i.e.  "grants management"  get's "." instead of space due to the make.names argument in dataframe() being TRUE by default. The workaround was to double sub   " " -> "_" then search using "_" and then finally, present the data as " " again.